Add Input Hyperslabing & Generalize HDF Dimensions
Achieved parallel input via HDF hyperslabs using the existing MPI
processes. Changed the HDF structure that `txt2hdf.c` outputs, and
thus changed how the HDF file is read in `init.c`. Structure of hdf
input needs to be changed to Dream3D standard, as that will be the
primary interface. Also updated how the arrays for the HDF data are
created in memory, using dimensions from `input.in` instead of hardcoding
- these values are also used for the hyperslab parameters.