Add Dynamic HDF Path Support
Added support for changing the paths to the euler angle, grain ID, and
phase ID datasets inside the HDF file by way of the `input.in` file
passed in. This required some structural changes to how `io.c` takes in
data for the NameList, particularly concerning string input, and strings
now must only be delimited with a space between the token and the
argument.