Fix Hyperslabbing Issues and Code Cleanup
Fixed the problem where the dataset wasn't being hyperslabbed properly.
This involved creating a dataspace within memory to match that of the
dataset. Also did some code cleanup to reduce warnings. The `.dream3d`
files appear to use a 4 dimensional array to describe 3 dimensional
grain id/phase id data (the 4th dimension size just being 1), so the
memory dataspace now accomodates for that. This appears to work for both
datasets that have 4 dimensions but only 1 item in the 4th as well as
standard 3 dimensional datasets.