Add Parallel HDF Installation Support
Added HDF5 source code to repo and modified the `installer.sh` script to use that
and install hdf5 with paralell support locally inside the repository. Modified
the Makefile to use the locally installed version of hdf5 instead of looking for
a global one. This should make installation of M-THEMES easier.